+++ wire-tests-swift/no-manifest/src/main/swift/MessageWithStatus.swift
@@ -75,7 +75,7 @@ extension MessageWithStatus : Codable {
  */
 extension MessageWithStatus {
 
-    public enum Status : Swift.UInt32, Swift.CaseIterable, Wire.ProtoEnum {
+    public enum Status : Swift.Int32, Swift.CaseIterable, Wire.ProtoEnum {
 
         case A = 1