+++ wire-tests-swift/no-manifest/src/main/swift/OptionalEnumUser.swift
@@ -103,7 +103,7 @@ extension OptionalEnumUser : Codable {
  */
 extension OptionalEnumUser {
 
-    public enum OptionalEnum : Swift.UInt32, Swift.CaseIterable, Wire.ProtoEnum {
+    public enum OptionalEnum : Swift.Int32, Swift.CaseIterable, Wire.ProtoEnum {
 
         case FOO = 1
         case BAR = 2