+++ CHANGELOG.md
@@ -13,6 +13,8 @@ _Unreleased_
     * This is attached to optional scalar values and messages with entirely optional values
 * New: `ProtoDefaulted` and `CustomDefaulted` include setter support
     * This enables you to do something like `Foo().$bar.$baz += 1`
+* Change: Swift `ProtoEnum` types now have a raw value of `Int32`.
+    * The runtime and generated code both need to be updated to reflect this.
 
 Version 4.9.1
 -------------